Upgrade deps

+++ src/main/java/vlaship/backoffice/repository/CategoryRepository.java
@@ -6,7 +6,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface CategoryRepository extends JpaRepository<Category, Integer> {
+public interface CategoryRepository extends JpaRepository<Category, Long> {
 
     List<Category> findAllByName(String name, Pageable pageable);
+
 }
